moved location examples to new package
