+++ MapboxAndroidDemo/src/main/AndroidManifest.xml
@@ -149,14 +149,14 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".examples.plugins.KotlinLocationComponentActivity"
+            android:name=".examples.location.KotlinLocationComponentActivity"
             android:label="@string/activity_location_location_component_title">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".examples.plugins.LocationComponentActivity"
+            android:name=".examples.location.LocationComponentActivity"
             android:label="@string/activity_location_location_component_title">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
@@ -170,7 +170,7 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".examples.plugins.LocationComponentFragmentActivity"
+            android:name=".examples.location.LocationComponentFragmentActivity"
             android:label="@string/activity_location_user_location_map_frag_title">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
