+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/examples/location/KotlinLocationComponentActivity.kt
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxandroiddemo.examples.plugins
+package com.mapbox.mapboxandroiddemo.examples.location
 
 import android.annotation.SuppressLint
 import android.os.Bundle
